requirements.txt
docopt==0.6.2
Flask==0.12.2
progressbar2==3.30.2
requests==2.18.1
xmltodict==0.11.0
